11/requirements.txt
Django~=3.1.7
djangorestframework~=3.12.4
graphene_django~=2.15.0
django-graphql-jwt~=0.3.0
django-graphql-auth~=0.3.15
django-filter~=2.4.0
PyJWT==1.7.0
mock~=4.0.3
drf-yasg~=1.20.0